+++ src/pydvl/value/shapley/gt.py
@@ -16,17 +16,20 @@ You can read more :ref:`in the documentation<data valuation>`.
 
 """
 import logging
+import numbers
 from collections import namedtuple
-from typing import Iterable, Tuple, TypeVar, cast
+from typing import Iterable, Optional, Tuple, TypeVar, cast
 
 import cvxpy as cp
 import numpy as np
+from numpy.random import SeedSequence
 from numpy.typing import NDArray
 
 from pydvl.utils import MapReduceJob, ParallelConfig, Utility, maybe_progress
 from pydvl.utils.numeric import random_subset_of_size
 from pydvl.utils.parallel.backend import effective_n_jobs
 from pydvl.utils.status import Status
+from pydvl.utils.types import Seed, SeedOrGenerator
 from pydvl.value import ValuationResult
 
 __all__ = ["group_testing_shapley", "num_samples_eps_delta"]
@@ -114,7 +117,11 @@ def num_samples_eps_delta(
 
 
 def _group_testing_shapley(
-    u: Utility, n_samples: int, progress: bool = False, job_id: int = 1
+    u: Utility,
+    n_samples: int,
+    progress: bool = False,
+    job_id: int = 1,
+    seed: Seed = None,
 ):
     """Helper function for :func:`group_testing_shapley`.
 
@@ -125,9 +132,10 @@ def _group_testing_shapley(
     :param n_samples: total number of samples (subsets) to use.
     :param progress: Whether to display progress bars for each job.
     :param job_id: id to use for reporting progress (e.g. to place progres bars)
+    :param seed: Seed for the random number generator.
     :return:
     """
-    rng = np.random.default_rng()
+    rng = np.random.default_rng(seed)
     n = len(u.data.indices)
     const = _constants(n, 1, 1, 1)  # don't care about eps,delta,range
 
@@ -138,7 +146,7 @@ def _group_testing_shapley(
 
     for t in maybe_progress(n_samples, progress=progress, position=job_id):
         k = rng.choice(const.kk, size=1, p=const.q).item()
-        s = random_subset_of_size(u.data.indices, k)
+        s = random_subset_of_size(u.data.indices, k, seed=rng)
         uu[t] = u(s)
         betas[t, s] = 1
     return uu, betas
@@ -153,6 +161,7 @@ def group_testing_shapley(
     n_jobs: int = 1,
     config: ParallelConfig = ParallelConfig(),
     progress: bool = False,
+    seed: Seed = None,
     **options,
 ) -> ValuationResult:
     """Implements group testing for approximation of Shapley values as described
@@ -181,6 +190,7 @@ def group_testing_shapley(
     :param config: Object configuring parallel computation, with cluster
         address, number of cpus, etc.
     :param progress: Whether to display progress bars for each job.
+    :param seed: Seed for the random number generator.
     :param options: Additional options to pass to `cvxpy.Problem.solve()
         <https://www.cvxpy.org/tutorial/advanced/index.html#solve-method-options>`_.
         E.g. to change the solver (which defaults to `cvxpy.SCS`) pass
@@ -226,6 +236,7 @@ def group_testing_shapley(
         map_kwargs=dict(n_samples=samples_per_job, progress=progress),
         config=config,
         n_jobs=n_jobs,
+        seed=seed,
     )
     uu, betas = map_reduce_job()