+++ src/pydvl/value/shapley/owen.py
@@ -1,14 +1,17 @@
+import numbers
 import operator
 from enum import Enum
 from functools import reduce
 from itertools import cycle, takewhile
-from typing import Sequence
+from typing import Optional, Sequence
 
 import numpy as np
+from numpy.random import SeedSequence
 from numpy.typing import NDArray
 from tqdm import tqdm
 
 from pydvl.utils import MapReduceJob, ParallelConfig, Utility, random_powerset
+from pydvl.utils.types import Seed, SeedOrGenerator
 from pydvl.value import ValuationResult
 from pydvl.value.stopping import MinUpdates
 
@@ -29,6 +32,7 @@ def _owen_sampling_shapley(
     *,
     progress: bool = False,
     job_id: int = 1,
+    seed: Seed = None,
 ) -> ValuationResult:
     r"""This is the algorithm as detailed in the paper: to compute the outer
     integral over q ∈ [0,1], use uniformly distributed points for evaluation
@@ -47,8 +51,10 @@ def _owen_sampling_shapley(
     :param max_q: number of subdivisions for the integration over $q$
     :param progress: Whether to display progress bars for each job
     :param job_id: For positioning of the progress bar
+    :param seed: Seed for the random number generator.
     :return: Object with the data values, errors.
     """
+    rng = np.random.default_rng(seed)
     q_stop = {OwenAlgorithm.Standard: 1.0, OwenAlgorithm.Antithetic: 0.5}
     q_steps = np.linspace(start=0, stop=q_stop[method], num=max_q)
 
@@ -67,7 +73,7 @@ def _owen_sampling_shapley(
         e = np.zeros(max_q)
         subset = np.setxor1d(u.data.indices, [idx], assume_unique=True)
         for j, q in enumerate(q_steps):
-            for s in random_powerset(subset, n_samples=n_samples, q=q):
+            for s in random_powerset(subset, n_samples=n_samples, q=q, seed=rng):
                 marginal = u({idx}.union(s)) - u(s)
                 if method == OwenAlgorithm.Antithetic and q != 0.5:
                     s_complement = np.setxor1d(subset, s, assume_unique=True)
@@ -93,6 +99,7 @@ def owen_sampling_shapley(
     n_jobs: int = 1,
     config: ParallelConfig = ParallelConfig(),
     progress: bool = False,
+    seed: Seed = None,
 ) -> ValuationResult:
     r"""Owen sampling of Shapley values as described in
     :footcite:t:`okhrati_multilinear_2021`.
@@ -138,6 +145,7 @@ def owen_sampling_shapley(
     :param config: Object configuring parallel computation, with cluster
         address, number of cpus, etc.
     :param progress: Whether to display progress bars for each job.
+    :param seed: Seed for the random number generator.
     :return: Object with the data values.
 
     .. versionadded:: 0.3.0
@@ -159,6 +167,7 @@ def owen_sampling_shapley(
         ),
         n_jobs=n_jobs,
         config=config,
+        seed=seed,
     )
 
     return map_reduce_job()